refactor(api): move assessment pre-handler to shared
