+++ api/src/evaluation/application/assessments/index.js
@@ -4,7 +4,6 @@ import { config } from '../../../../config/config.js';
 import { checkLLMChatIsEnabled } from '../../../llm/application/pre-handlers/index.js';
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
-import { assessmentAuthorization } from '../pre-handlers/assessment-authorization.js';
 import { assessmentController } from './assessment-controller.js';
 
 const { featureToggles } = config;
@@ -23,7 +22,7 @@ const register = async function (server) {
         },
         pre: [
           {
-            method: assessmentAuthorization.verify,
+            method: securityPreHandlers.checkUserOwnsAssessment,
             assign: 'authorizationCheck',
           },
         ],
@@ -38,7 +37,7 @@ const register = async function (server) {
         auth: false,
         pre: [
           {
-            method: assessmentAuthorization.verify,
+            method: securityPreHandlers.checkUserOwnsAssessment,
             assign: 'authorizationCheck',
           },
         ],
@@ -136,7 +135,7 @@ const register = async function (server) {
         auth: false,
         pre: [
           {
-            method: assessmentAuthorization.verify,
+            method: securityPreHandlers.checkUserOwnsAssessment,
             assign: 'authorizationCheck',
           },
         ],
