+++ api/tests/shared/unit/application/security-pre-handlers_test.js
@@ -1489,4 +1489,58 @@ describe('Shared | Unit | Application | SecurityPreHandlers', function () {
       });
     });
   });
+
+  describe('#checkUserOwnsAssessment', function () {
+    let assessmentRepository;
+    let validationErrorSerializer;
+
+    beforeEach(function () {
+      assessmentRepository = { getByAssessmentIdAndUserId: sinon.stub() };
+      validationErrorSerializer = { serialize: sinon.stub() };
+    });
+
+    describe('when user is the owner of the assessment', function () {
+      it('should return the assessment', async function () {
+        const request = { auth: { credentials: { userId: 100 } }, params: { id: 8 } };
+        assessmentRepository.getByAssessmentIdAndUserId.resolves({});
+
+        const response = await securityPreHandlers.checkUserOwnsAssessment(request, hFake, {
+          assessmentRepository,
+          validationErrorSerializer,
+        });
+
+        sinon.assert.calledWith(assessmentRepository.getByAssessmentIdAndUserId, 8, 100);
+        expect(response).to.deep.equal({});
+      });
+    });
+
+    describe('when the assessment has no owner', function () {
+      it('should return the assessment', async function () {
+        const request = { auth: { credentials: null }, params: { id: 8 } };
+        assessmentRepository.getByAssessmentIdAndUserId.resolves({});
+
+        const response = await securityPreHandlers.checkUserOwnsAssessment(request, hFake, {
+          assessmentRepository,
+          validationErrorSerializer,
+        });
+
+        sinon.assert.calledWith(assessmentRepository.getByAssessmentIdAndUserId, 8, undefined);
+        expect(response).to.deep.equal({});
+      });
+    });
+
+    describe('when user is not the owner of the assessment', function () {
+      it('should return a status 401', async function () {
+        const request = { auth: { credentials: { userId: 101 } }, params: { id: 8 } };
+        assessmentRepository.getByAssessmentIdAndUserId.rejects();
+
+        const response = await securityPreHandlers.checkUserOwnsAssessment(request, hFake, {
+          assessmentRepository,
+          validationErrorSerializer,
+        });
+
+        expect(response.statusCode).to.equal(401);
+      });
+    });
+  });
 });
